Use the SageBuilder interface instead of calling constructors.
Running the cuda translator with Valgrind revealed a usage of
unintialized value when building AST objects. It turned out that
it is caused by InitializedName objects created directly with its
constructor rather than the SageBuilder interface. The previous
way worked correctly but it is certainly an undesirable way to create
its instances, and may manifest itself as unwanted behavior in
different locations.