Add the reduction intrinsic to Translator.
The intrinsic is now recognized by TranslationContext, and its use is
attached with an attribute for annotating reduction-related information. The
abstract base class, Translator, has a do-nothing implementation for reduction.
No concrete classes (e.g., ReferenceTraslator) yet have implementation, so
any use of the intrisic appears in output code as it is written in input code.