Fix function creation with parameter references
Function parameters may not be used in the function body if they are not
yet built and inserted by buildDefiningFunction and insertStatement.
Otherwise, the parameters are not actually visible inside the function
body, so buildVarRefExp will fail to create actual references, and
result in dummy placeholder references with type SgUnknownType. This
will in turn cause SageInterface::clearUnusedVariables to delete those
dummy variable symbols since they are of SgUnknownType.

This did not manifest as a problem ROSE versions up to Git revision
26dd7c8f. The commit c219a5e added a call to clearUnusedVariables in
SageInterface::fixVariableReferences, which is called inside
SageBuilder::buildForStatement. More specifically, in
cuda_runtime_builder.cc, a function to create a grid object of a
user-specified type is created using buildDefiningFunction. Later, a
functions for copyin and copyout are also created, which have a for
statement inside. Since the former function used to create parameter
references before building its function object itself, they resulted in
being deleted when the for statement was created.