Fix missing scope of variables (edg4x)
SgInitializedName of a Variable built with SageBuilder::buildVariableDeclaration
with scope parameter NULL was automatically assigned a proper scope when the
variable declaration statement is assigned to a basic block. It does not
work so with ROSE-Edg4x, but the scope is set to SgGlobal. Explicitly
passing the proper scope variable solves the problem.