Set a proper scope when building a variable declaration (edg4x).
The scope of a symbol generated by buildVariableDeclaration used to be
set when the declaration statement is added to a scope, such as
SageInterface::insertStatement(). It does not seem to be the case with
ROSE-Edg4x. The scope of variable declaration statements is set when
it's added to a scope, but the symbol of the variable is not. This
actually caused a problem (see commit f85f28c). Although no problem
shows up anymore, it would be more robust to change all the uses of
buildVariableDeclaration to have a non-NULL scope argument.